// back-end/src/play/play.interface.ts
export class Play {
  PLAY_SQ: number;
  PLAY_NM: string;
  PLAY_STDD: string;
  PLAY_DESC: string;
  PLAY_EPNTN: string;
  PLAY_PARENTS_MSG: string;
  PLAY_FEAT: string;
  PLAY_STRONG: string;
  PLAY_WEAK: string;
  PLAY_HOME: string;
  PLAY_PARENTS: string;
  PLAY_GUIDE: string;
  PLAY_SPORTS: string;
  PLAY_LEARN: string;
  CRT_DT: string;
  MOD_DT: string;
  DEL_DT: string;
  PLAY_PNT: number;
  PLAY_IMG_PATH: string;
  PLAY_HOME_IMG_PATH: string;
  PLAY_PARENTS_IMG_PATH: string;
  PLAY_GUIDE_IMG_PATH: string;
  PLAY_SPORTS_IMG_PATH: string;
  PLAY_LEARN_IMG_PATH: string;
}

export class PlayStandard {
  PLAY_STDD_SQ: number;
  PLAY_SQ: number;
  TEST_CTGR_SQ: number;
}